move factories and datasource's in own module
(cherry picked from commit 42dc9be4373794415b0d6dd779cd919ac829058b)
