+++ self_description_service.factories/src/main/java/org/eclipse/slm/self_description_service/factories/TemplateDatasource.java
@@ -1,4 +1,4 @@
-package org.eclipse.slm.self_description_service.aas.factories;
+package org.eclipse.slm.self_description_service.factories;
 
 import org.apache.poi.openxml4j.exceptions.InvalidFormatException;
 import org.eclipse.digitaltwin.aas4j.v3.dataformat.aasx.AASXDeserializer;
@@ -7,9 +7,8 @@ import org.eclipse.digitaltwin.aas4j.v3.model.Environment;
 import org.eclipse.digitaltwin.aas4j.v3.model.Property;
 import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
 import org.eclipse.digitaltwin.aas4j.v3.model.SubmodelElement;
-import org.eclipse.slm.self_description_service.ITemplateManager;
-import org.eclipse.slm.self_description_service.TemplateManager;
-import org.eclipse.slm.self_description_service.aas.Datasource;
+import org.eclipse.slm.self_description_service.factories.template.ITemplateManager;
+import org.eclipse.slm.self_description_service.factories.template.TemplateManager;
 import org.eclipse.slm.self_description_service.templating.TemplateRenderer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
