+++ self_description_service.factories/src/main/java/org/eclipse/slm/self_description_service/factories/template/TemplateManager.java
@@ -1,4 +1,4 @@
-package org.eclipse.slm.self_description_service;
+package org.eclipse.slm.self_description_service.factories.template;
 
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
