+++ self_description_service.factories/src/test/resources/templates/Template.aasx
